chore: convert ThemeToggleProvider and Theme
