+++ packages/core/admin/admin/custom.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+import { StrapiTheme } from '@strapi/design-system';
+
+declare module 'styled-components' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  export interface DefaultTheme extends StrapiTheme {}
+}
