--- packages/core/admin/admin/src/components/GlobalStyle/index.js
@@ -1,9 +1,0 @@
-import { createGlobalStyle } from 'styled-components';
-
-const GlobalStyle = createGlobalStyle`
-  body {
-    background: ${({ theme }) => theme.colors.neutral100};
-  }
-`;
-
-export default GlobalStyle;
