+++ packages/core/admin/admin/src/content-manager/pages/EditView/Header/tests/index.test.js
@@ -11,8 +11,8 @@ import { render } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';
 
-import Theme from '../../../../../components/Theme';
-import ThemeToggleProvider from '../../../../../components/ThemeToggleProvider';
+import { Theme } from '../../../../../components/Theme';
+import { ThemeToggleProvider } from '../../../../../components/ThemeToggleProvider';
 import { Header } from '../index';
 
 import ct from './data/ct-schema.json';
