+++ packages/core/admin/admin/src/pages/SettingsPage/pages/ApiTokens/ListView/tests/index.test.js
@@ -10,8 +10,8 @@ import { Provider } from 'react-redux';
 import { Route, MemoryRouter } from 'react-router-dom';
 import { createStore } from 'redux';
 
-import Theme from '../../../../../../components/Theme';
-import ThemeToggleProvider from '../../../../../../components/ThemeToggleProvider';
+import { Theme } from '../../../../../../components/Theme';
+import { ThemeToggleProvider } from '../../../../../../components/ThemeToggleProvider';
 import ListView from '../index';
 
 jest.mock('@strapi/helper-plugin', () => ({
