+++ packages/core/admin/admin/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "tsconfig/client.json",
-  "include": ["src"],
+  "include": ["src", "custom.d.ts"],
   "exclude": ["node_modules"]
 }
